+++ packages/seahorse/src/primitives/host/index.tsx
@@ -1,0 +1,2 @@
+export { ForgeHost as Host } from '../_host/ForgeHost'
+export type { UniversalHostProps as HostProps } from '@expo/ui'