+++ packages/seahorse/src/primitives/notification-dot/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Box } from '../box'
-import { Text } from '../text'
+import { Text } from 'react-native-css/components'
 import { cn } from '../../utils/cn'
 
 type NotificationDotProps = {